优化用户不能删除自己

+++ aidex-controller/src/main/java/com/aidex/web/controller/system/SysUserController.java
@@ -21,6 +21,7 @@ import com.aidex.system.service.ISysUserService;
 import com.aidex.system.service.SysPostService;
 import com.alibaba.fastjson.JSONObject;
 import com.github.pagehelper.PageInfo;
+import org.apache.commons.lang3.ArrayUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.validation.annotation.Validated;
@@ -36,13 +37,12 @@ import java.util.stream.Collectors;
 
 /**
  * 用户信息
- * 
+ *
  * @author ruoyi
  */
 @RestController
 @RequestMapping("/system/user")
-public class SysUserController extends BaseController
-{
+public class SysUserController extends BaseController {
     @Autowired
     private ISysUserService userService;
 
@@ -60,8 +60,7 @@ public class SysUserController extends BaseController
      */
     @PreAuthorize("@ss.hasPermi('system:user:list')")
     @GetMapping("/list")
-    public R<PageInfo> page(SysUser user, HttpServletRequest request, HttpServletResponse response)
-    {
+    public R<PageInfo> page(SysUser user, HttpServletRequest request, HttpServletResponse response) {
         user.setPage(new PageDomain(request, response));
         return R.data(userService.findPage(user));
     }
@@ -69,8 +68,7 @@ public class SysUserController extends BaseController
     @Log(title = "用户管理", businessType = BusinessType.EXPORT)
     @PreAuthorize("@ss.hasPermi('system:user:export')")
     @GetMapping("/export")
-    public R export(SysUser user)
-    {
+    public R export(SysUser user) {
         List<SysUser> list = userService.findList(user);
         ExcelUtil<SysUser> util = new ExcelUtil<SysUser>(SysUser.class);
         return util.exportExcel(list, "用户数据");
@@ -79,8 +77,7 @@ public class SysUserController extends BaseController
     @Log(title = "用户管理", businessType = BusinessType.IMPORT)
     @PreAuthorize("@ss.hasPermi('system:user:import')")
     @PostMapping("/importData")
-    public AjaxResult importData(MultipartFile file, boolean updateSupport) throws Exception
-    {
+    public AjaxResult importData(MultipartFile file, boolean updateSupport) throws Exception {
         ExcelUtil<SysUser> util = new ExcelUtil<SysUser>(SysUser.class);
         List<SysUser> userList = util.importExcel(file.getInputStream());
         LoginUser loginUser = tokenService.getLoginUser(ServletUtils.getRequest());
@@ -90,8 +87,7 @@ public class SysUserController extends BaseController
     }
 
     @GetMapping("/importTemplate")
-    public R importTemplate()
-    {
+    public R importTemplate() {
         ExcelUtil<SysUser> util = new ExcelUtil<SysUser>(SysUser.class);
         return util.importTemplateExcel("用户数据");
     }
@@ -100,15 +96,13 @@ public class SysUserController extends BaseController
      * 根据用户编号获取详细信息
      */
     @PreAuthorize("@ss.hasPermi('system:user:query')")
-    @GetMapping(value = { "/", "/{userId}" })
-    public AjaxResult getInfo(@PathVariable(value = "userId", required = false) String userId)
-    {
+    @GetMapping(value = {"/", "/{userId}"})
+    public AjaxResult getInfo(@PathVariable(value = "userId", required = false) String userId) {
         AjaxResult ajax = AjaxResult.success();
         List<SysRole> roles = roleService.selectRoleAll();
         ajax.put("roles", SysUser.isAdmin(userId) ? roles : roles.stream().filter(r -> !r.isAdmin()).collect(Collectors.toList()));
         ajax.put("posts", postService.findList(new SysPost()));
-        if (StringUtils.isNotNull(userId))
-        {
+        if (StringUtils.isNotNull(userId)) {
             ajax.put(AjaxResult.DATA_TAG, userService.get(userId));
             ajax.put("postIds", postService.selectPostListByUserId(userId));
             ajax.put("roleIds", roleService.selectRoleListByUserId(userId));
@@ -122,8 +116,7 @@ public class SysUserController extends BaseController
     @PreAuthorize("@ss.hasPermi('system:user:add')")
     @Log(title = "用户管理", businessType = BusinessType.INSERT)
     @PostMapping
-    public R add(@Validated @RequestBody SysUser user)
-    {
+    public R add(@Validated @RequestBody SysUser user) {
         return R.status(userService.insertUser(user));
     }
 
@@ -133,8 +126,7 @@ public class SysUserController extends BaseController
     @PreAuthorize("@ss.hasPermi('system:user:edit')")
     @Log(title = "用户管理", businessType = BusinessType.UPDATE)
     @PutMapping
-    public R edit(@Validated @RequestBody SysUser user)
-    {
+    public R edit(@Validated @RequestBody SysUser user) {
         userService.checkUserAllowed(user);
         return R.status(userService.updateUser(user));
     }
@@ -145,8 +137,10 @@ public class SysUserController extends BaseController
     @PreAuthorize("@ss.hasPermi('system:user:remove')")
     @Log(title = "用户管理", businessType = BusinessType.DELETE)
     @DeleteMapping("/{userIds}")
-    public R remove(@PathVariable("userIds") String[] userIds)
-    {
+    public R remove(@PathVariable("userIds") String[] userIds) {
+        if (ArrayUtils.contains(userIds, getUserId())) {
+            return R.fail("当前用户不能删除");
+        }
         return R.status(userService.deleteUserByIds(userIds));
     }
 
@@ -156,8 +150,7 @@ public class SysUserController extends BaseController
     @PreAuthorize("@ss.hasPermi('system:user:resetPwd')")
     @Log(title = "用户管理", businessType = BusinessType.UPDATE)
     @PutMapping("/resetPwd")
-    public R resetPwd(@RequestBody SysUser user)
-    {
+    public R resetPwd(@RequestBody SysUser user) {
         userService.checkUserAllowed(user);
         user.setPassword(SecurityUtils.encryptPassword(user.getPassword()));
         user.setUpdateBy(SecurityUtils.getUsername());
@@ -170,12 +163,12 @@ public class SysUserController extends BaseController
     @PreAuthorize("@ss.hasPermi('system:user:edit')")
     @Log(title = "用户管理", businessType = BusinessType.UPDATE)
     @PutMapping("/changeStatus")
-    public R changeStatus(@RequestBody SysUser user)
-    {
+    public R changeStatus(@RequestBody SysUser user) {
         userService.checkUserAllowed(user);
         user.setUpdateBy(SecurityUtils.getUsername());
         return R.status(userService.updateUserStatus(user));
     }
+
     /**
      * 校验角色编码是否存在
      */
@@ -183,15 +176,16 @@ public class SysUserController extends BaseController
     @Log(title = "用户管理", businessType = BusinessType.CHECK)
     @GetMapping("/checkUserNameUnique")
     public R checkUserNameUnique(SysUser user) {
-        Map<String,String> checkMap = new HashMap<String,String>(1);
-        try{
+        Map<String, String> checkMap = new HashMap<String, String>(1);
+        try {
             userService.checkUserNameUnique(user);
-            checkMap.put("code","1");
-        }catch (Exception e){
-            checkMap.put("code","2");
+            checkMap.put("code", "1");
+        } catch (Exception e) {
+            checkMap.put("code", "2");
         }
         return R.data(checkMap);
     }
+
     /**
      * 校验角色编码是否存在
      */
@@ -199,15 +193,16 @@ public class SysUserController extends BaseController
     @Log(title = "用户管理", businessType = BusinessType.CHECK)
     @GetMapping("/checkEmailUnique")
     public R checkEmailUnique(SysUser user) {
-        Map<String,String> checkMap = new HashMap<String,String>(1);
-        try{
+        Map<String, String> checkMap = new HashMap<String, String>(1);
+        try {
             userService.checkEmailUnique(user);
-            checkMap.put("code","1");
-        }catch (Exception e){
-            checkMap.put("code","2");
+            checkMap.put("code", "1");
+        } catch (Exception e) {
+            checkMap.put("code", "2");
         }
         return R.data(checkMap);
     }
+
     /**
      * 校验角色编码是否存在
      */
@@ -215,29 +210,30 @@ public class SysUserController extends BaseController
     @Log(title = "用户管理", businessType = BusinessType.CHECK)
     @GetMapping("/checkPhoneUnique")
     public R checkPhoneUnique(SysUser user) {
-        Map<String,String> checkMap = new HashMap<String,String>(1);
-        try{
+        Map<String, String> checkMap = new HashMap<String, String>(1);
+        try {
             userService.checkPhoneUnique(user);
-            checkMap.put("code","1");
-        }catch (Exception e){
-            checkMap.put("code","2");
+            checkMap.put("code", "1");
+        } catch (Exception e) {
+            checkMap.put("code", "2");
         }
         return R.data(checkMap);
     }
 
     /**
      * 通用选人页面根据用户ID
+     *
      * @return
      */
-    @PostMapping(value = { "/getUserInfoByIds"})
-    public R getUserInfoByIds(@Validated @RequestBody JSONObject userIdsObj)
-    {
-        List<Map<String,Object>> sysUserList = userService.getUserInfoByIds(userIdsObj);
+    @PostMapping(value = {"/getUserInfoByIds"})
+    public R getUserInfoByIds(@Validated @RequestBody JSONObject userIdsObj) {
+        List<Map<String, Object>> sysUserList = userService.getUserInfoByIds(userIdsObj);
         return R.data(sysUserList);
     }
 
     /**
      * 获取角色下所有用户信息
+     *
      * @param sysUser
      * @param request
      * @param response
@@ -245,8 +241,7 @@ public class SysUserController extends BaseController
      */
     @PreAuthorize("@ss.hasPermi('system:user:list')")
     @GetMapping("/roleUserList")
-    public R<PageInfo> roleUserPage(SysUser sysUser, HttpServletRequest request, HttpServletResponse response)
-    {
+    public R<PageInfo> roleUserPage(SysUser sysUser, HttpServletRequest request, HttpServletResponse response) {
         sysUser.setPage(new PageDomain(request, response));
         return R.data(userService.findRoleUserPage(sysUser));
     }
@@ -254,8 +249,7 @@ public class SysUserController extends BaseController
     @PreAuthorize("@ss.hasPermi('system:user:add')")
     @Log(title = "用户角色管理", businessType = BusinessType.INSERT)
     @PostMapping("/addRoleUser")
-    public R addRoleUser(@Validated @RequestBody SysUserRole sysUserRole)
-    {
+    public R addRoleUser(@Validated @RequestBody SysUserRole sysUserRole) {
         return R.status(userService.insertRoleUser(sysUserRole));
     }
 }
